+++ cmn-data-svc-client/src/main/java/org/acumos/cds/CCDSConstants.java
@@ -92,6 +92,9 @@ public interface CCDSConstants {
 	public static final String LOGIN_API_PATH = "loginapi";
 	public static final String VERIFY_PATH = "verify";
 	public static final String CHPASS_PATH = "chgpw";
+	
+	//Actions on datasets
+	public static final String DATASET_PATH = "dataset";
 
 	// Portal search interface
 	public static final String SEARCH_ACCESS_TYPES = "atc";